print basic invoices

+++ controllers/admin.js
@@ -3,6 +3,7 @@
 const { ValidationError } = require('sequelize')
 const Product = require('../models/product')
 const { validationResult } = require('express-validator')
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
 	res.render('admin/edit-product', {
@@ -17,9 +18,25 @@ exports.getAddProduct = (req, res, next) => {
 }
 exports.postAddProduct = async (req, res, next) => {
 	const title = req.body.title
-	const imageUrl = req.body.imageUrl
+	const image = req.file
 	const price = req.body.price
 	const description = req.body.description
+
+
+	if (!image) {
+		return res.status(422).render('admin/edit-product', {
+			pageTitle: 'Add Product',
+			path: '/admin/edit-product',
+			editing: false,
+			hasError: true,
+			product: { title, price, description },
+			errorMsg: 'Attached file is not an image',
+			validationErrors: []
+
+		})
+	}
+	const imageUrl = image.path
+
 	const errors = validationResult(req)
 	if (!errors.isEmpty()) {
 		return res.status(422).render('admin/edit-product', {
@@ -27,7 +44,7 @@ exports.postAddProduct = async (req, res, next) => {
 			path: '/admin/edit-product',
 			editing: false,
 			hasError: true,
-			product: { title, imageUrl, price, description },
+			product: { title, price, description },
 			errorMsg: errors.array()[0].msg,
 			validationErrors: errors.array()
 
@@ -38,7 +55,9 @@ exports.postAddProduct = async (req, res, next) => {
 		await product.save()
 		res.redirect('/admin/products')
 	} catch (err) {
-		console.log(err)
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 exports.getEditProduct = async (req, res, next) => {
@@ -62,8 +81,10 @@ exports.getEditProduct = async (req, res, next) => {
 			validationErrors: []
 
 		})
-	} catch (error) {
-		console.log(error)
+	} catch (err) {
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 
@@ -71,8 +92,11 @@ exports.postEditProduct = async (req, res, next) => {
 	const prodId = req.body.productId
 	const updatedTitle = req.body.title
 	const updatedPrice = req.body.price
-	const updatedImageUrl = req.body.imageUrl
+	const image = req.file
 	const updatedDescription = req.body.description
+
+
+
 
 	const errors = validationResult(req)
 	if (!errors.isEmpty()) {
@@ -81,7 +105,7 @@ exports.postEditProduct = async (req, res, next) => {
 			path: '/admin/edit-product',
 			editing: true,
 			hasError: true,
-			product: { title: updatedTitle, imageUrl: updatedImageUrl, price: updatedPrice, description: updatedDescription },
+			product: { title: updatedTitle, price: updatedPrice, description: updatedDescription },
 			errorMsg: errors.array()[0].msg,
 			validationErrors: errors.array(),
 			_id: prodId
@@ -94,12 +118,17 @@ exports.postEditProduct = async (req, res, next) => {
 		}
 		newProduct.title = updatedTitle
 		newProduct.price = updatedPrice
-		newProduct.imageUrl = updatedImageUrl
+		if (image) {
+			fileHelper.deleteFile(product.imageUrl)
+			newProduct.imageUrl = image.path
+		}
 		newProduct.description = updatedDescription
 		res.redirect('/admin/products')
 		return await newProduct.save()
-	} catch (error) {
-		console.log(error)
+	} catch (err) {
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 exports.getProducts = async (req, res, next) => {
@@ -110,16 +139,27 @@ exports.getProducts = async (req, res, next) => {
 			pageTitle: 'Admin Products',
 			path: '/admin/products',
 		})
-	} catch (error) {
-		console.log(error)
+	} catch (err) {
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
 exports.postDeleteProduct = async (req, res, next) => {
 	const prodId = req.body.productId
+	
 	try {
+		const product = await Product.findById(prodId)
+		if(!product){
+			return next(new Error('Product not found.'))
+		}
+		fileHelper.deleteFile(product.imageUrl)
 		await Product.deleteOne({ _id: prodId, userId: req.user._id })
+		
 		res.redirect('/admin/products')
-	} catch (error) {
-		console.log(error)
+	} catch (err) {
+		const error = new Error(err)
+		error.httpStatusCode = 500
+		return next(error)
 	}
 }
